Updated changelog to reflect recent commit changes.

+++ LUFA/Drivers/USB/LowLevel/Endpoint.c
@@ -83,7 +83,7 @@ void Endpoint_ClearEndpoints(void)
 
 uint8_t Endpoint_WaitUntilReady(void)
 {
-	uint8_t TimeoutMSRem = USB_STREAM_TIMEOUT_MS;
+	uint16_t TimeoutMSRem = USB_STREAM_TIMEOUT_MS;
 
 	USB_INT_Clear(USB_INT_SOFI);
